Move editor actions into a separate file
